BUG: Use the correct type of euler angles
Explicit is better than implicit, and the type of euler angle we were using was
both implicit and wrong.

This adds a template type that enforces that the convention is always explicit.

We previously used XYZ, now we use YXZ. See the comment in the diff for why. This demonstrably fixed a bug, visible by turning the robot upsidedown.